Refactor Token handler service
